Move implementations to separate files
