+++ syncademic_app/lib/authorization/authorization_service.dart
@@ -43,56 +43,4 @@ class MockAuthorizationService implements AuthorizationService {
   Future<String?> get userId => Future.value('mocked_user_id');
 }
 
-final _scopes = [CalendarApi.calendarScope];
-
-class GoogleAuthorizationService implements AuthorizationService {
-  final GoogleSignIn _googleSignIn;
-  GoogleSignInAccount? _currentUser;
-
-  GoogleAuthorizationService({GoogleSignIn? googleSignIn})
-      : _googleSignIn = googleSignIn ??
-            GoogleSignIn(
-              scopes: _scopes,
-              clientId: dotenv.env['SYNCADEMIC_CLIENT_ID'],
-            );
-
-  @override
-  Future<String?> getAuthorizationCode() async {
-    _currentUser = await _googleSignIn.signIn();
-    if (_currentUser == null) {
-      log('User not logged in');
-      return null;
-    }
-
-    return await requestServerAuthCode();
-  }
-
-  @override
-  Future<bool> authorize() async {
-    _currentUser = await _googleSignIn.signIn();
-    return _currentUser != null;
-  }
-
-  @override
-  Future<bool> isAuthorized() => _googleSignIn.canAccessScopes(_scopes);
-
-  @override
-  Future<String?> get accessToken async {
-    if (_currentUser == null) {
-      // TODO(SuperMuel) Log error
-      return null;
-    }
-
-    final headers = await _currentUser!.authHeaders;
-
-    // strip the "Bearer " prefix
-    return headers['Authorization']?.substring(7);
-  }
-
-  @override
-  Future<http.Client?> get authorizedClient =>
-      _googleSignIn.authenticatedClient();
-
-  @override
-  Future<String?> get userId async => _currentUser?.id;
-}
+final scopes = [CalendarApi.calendarScope];
