+++ syncademic_app/lib/authorization/backend_authorization_service.dart
@@ -1,37 +1,12 @@
 import 'dart:developer';
 
-import 'package:cloud_functions/cloud_functions.dart';
-import 'package:get_it/get_it.dart';
-
-import 'authorization_service.dart';
-
 abstract class BackendAuthorizationService {
   Future<void> authorizeBackend();
 }
 
-class FirebaseBackendAuthorizationService
-    implements BackendAuthorizationService {
+class MockBackendAuthorizationService implements BackendAuthorizationService {
   @override
   Future<void> authorizeBackend() async {
-    log("Authorizing the backend using Firebase");
-
-    final authCode =
-        await GetIt.I.get<AuthorizationService>().getAuthorizationCode();
-
-    if (authCode == null) {
-      log('Authorization code is null !');
-      throw Exception('Authorization code is null');
-    }
-
-    log("Got authorization code. Sending it to the backend.");
-
-    try {
-      await FirebaseFunctions.instance.httpsCallable('authorize_backend').call({
-        'authCode': authCode,
-      });
-    } on FirebaseFunctionsException catch (e) {
-      log('Error authorizing backend: ${e.code}, ${e.details}, ${e.message}');
-      rethrow;
-    }
+    log("Authorizing the backend using a mock service");
   }
 }
